Moved the CurrencyManager over to a new folder called Currency (in Core/Players/Currency). Created a new class called Currency that holds a currency's current values, meaning that now a currency class is stored instead of a dictionary. This allows new properties to be added to currencies much cleaner, as well as removes the need to remove and re-add a value to the currency table every time a value is changed.
