+++ Core/SubterfugeCore/Core/Players/Currency/Currency.cs
@@ -1,0 +1,8 @@
+﻿namespace SubterfugeCore.Core.Players.Currency
+{
+    class Currency
+    {
+        public int value = 0;
+        public bool? canBeNegative = true;
+    }
+}
